# packages/servers/MainServerGo/cmd/api/openapi.yaml
openapi: 3.0.3
info:
  title: DigiGold Platform API
  version: '1.1.0'
  description: |
    Multi-tenant digital gold platform. Jewelry shops (tenants) let their customers
    buy gold at live rates. Customers never sell: gold leaves a vault only as
    physical gold collected at the store counter (redemption with a pickup code).

    ## Conventions
    - **Base path:** `/api/v1`.
    - **Tenant:** every tenant-scoped route needs the `X-Tenant-ID` header (tenant UUID).
    - **Auth:** send the raw JWT in the `X-Api-Token` header. Do **not** add a `Bearer ` prefix.
    - **Token kinds:** customer access, customer refresh, registration, admin access and
      admin refresh tokens are not interchangeable. Each kind is rejected everywhere else.
    - **Access tokens** live 15 minutes, **refresh tokens** 7 days, **registration tokens** 10 minutes.
    - **IDs:** only UUIDs are exposed. Internal numeric IDs never appear in requests or responses.
    - **Rate limits:** `/auth/*` and `/admin/auth/*` allow 10 requests per minute per IP and route.
      Exceeding it returns `429`.
    - **Errors:** most routes return the `Error` schema. Some admin routes still return the
      older `SimpleError` shape (`{"error": "..."}`).
    - **5xx responses** never include internal details.
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
components:
  securitySchemes:
    CustomerToken:
      type: apiKey
      in: header
      name: X-Api-Token
      description: Customer access token from `/auth/otp/verify` or `/auth/register`. Raw JWT, no `Bearer` prefix.
    AdminToken:
      type: apiKey
      in: header
      name: X-Api-Token
      description: Admin access token from `/admin/auth/totp/verify` or `/admin/auth/refresh`. Raw JWT, no `Bearer` prefix.
  parameters:
    TenantId:
      name: X-Tenant-ID
      in: header
      required: true
      description: UUID of the tenant the request acts on.
      schema:
        type: string
        format: uuid
        example: 7e57d004-2b97-0e7a-b45f-5387367791cd
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: Invalid input.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired token, or invalid tenant.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Role not allowed, or token tenant does not match `X-Tenant-ID`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Entity not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooManyRequests:
      description: Rate limit, OTP cooldown or attempt limit reached.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message:
          {
            type: string,
            example: Insufficient gold balance for this transaction.,
          }
        code:
          type: integer
          description: App error code (see `interfaces/error-code.enum.go`).
          example: 244
        name: { type: string, example: ERROR_INSUFFICIENT_BALANCE }
        extra:
          description: Optional detail. Always empty on 5xx.
    SimpleError:
      type: object
      properties:
        error: { type: string }
...
    User:
      type: object
      properties:
        user_uuid: { type: string, format: uuid }
        full_name: { type: string, nullable: true }
        phone_number: { type: string }
        email_id: { type: string, nullable: true }
        kyc_status: { type: string, enum: [pending, verified, rejected] }
        document_json: { description: KYC document payload (JSON). }
        erp_unique_id: { type: string, nullable: true }
        total_vault_balance:
          { type: number, description: Gold balance in grams. }
        created_at: { type: string, format: date-time }
        modified_at: { type: string, format: date-time }
...
    LedgerEntry:
      type: object
      properties:
        gl_uuid: { type: string, format: uuid }
        event_type:
          {
            type: string,
            enum:
              [
                GOLD_PURCHASE,
                PHYSICAL_REDEMPTION,
                SYSTEM_REVERSAL,
                ADMIN_ADJUSTMENT,
              ],
          }
        payment_mode:
          { type: string, enum: [ONLINE_PG, COUNTER_CASH, COUNTER_UPI, NONE] }
        weight_grams:
          { type: number, description: 'Signed: + credit, - debit.' }
        total_amount_inr: { type: number }
        running_gold_balance_grams: { type: number }
        mcx_base_rate: { type: number }
        master_margin_applied: { type: number }
        tenant_margin_applied: { type: number }
        gst_applied: { type: number }
        final_rate_per_gram: { type: number }
        reference_id: { type: string }
        metadata_json: {}
        created_at: { type: string, format: date-time }
...
    TradeRequest:
      type: object
      properties:
        requested_rate_per_gram:
          type: number
          description: Rate shown to the user. Rejected if the live rate differs by more than ₹30.
        weight_grams:
          {
            type: number,
            description: Grams to buy. Takes priority over `total_amount_inr`.,
          }
        total_amount_inr: { type: number }
        payment_mode:
          { type: string, enum: [ONLINE_PG, COUNTER_CASH, COUNTER_UPI] }
        reference_id: { type: string }
    Paginated:
      type: object
      properties:
        success: { type: boolean }
        page: { type: integer }
        limit: { type: integer }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
paths:
...
  /admin/auth/login:
    post:
      operationId: postAdminAuthLogin
      tags: [Admin Auth]
      security: []
      summary: 'Step 1: password check'
      description: Returns a 5-minute `temp_token` for the TOTP step.
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        '200':
          description: Password accepted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  temp_token: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /admin/auth/totp/setup:
    post:
      operationId: postAdminAuthTotpSetup
      tags: [Admin Auth]
      security: []
      summary: 'Step 2a (first login only): enroll TOTP'
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [temp_token]
              properties:
                temp_token: { type: string }
      responses:
        '200':
          description: Show `otpauth_uri` as a QR code for an authenticator app.
          content:
            application/json:
              schema:
                type: object
                properties:
                  otpauth_uri: { type: string }
        '400':
          description: Invalid temp token or TOTP already enabled.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleError' }
  /admin/auth/totp/verify:
    post:
      operationId: postAdminAuthTotpVerify
      tags: [Admin Auth]
      security: []
      summary: 'Step 2b: verify TOTP code and get tokens'
      description: 5 wrong codes invalidate the temp token; log in again.
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [temp_token, code]
              properties:
                temp_token: { type: string }
                code: { type: string, example: '123456' }
      responses:
        '200':
          description: Logged in.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Invalid code or temp token.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleError' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
  /admin/auth/refresh:
    post:
      operationId: postAdminAuthRefresh
      tags: [Admin Auth]
      security: []
      summary: Exchange an admin refresh token for new tokens
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401':
          description: Invalid token or deactivated admin.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SimpleError' }

  # ------------------------------------------------------------------ Admin store
  /admin/store/customers:
    get:
      operationId: getAdminStoreCustomers
      tags: [Admin Store]
      summary: List tenant customers
      security: [{ AdminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Customers.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/store/ledger:
    get:
      operationId: getAdminStoreLedger
      tags: [Admin Store]
      summary: Tenant ledger
      security: [{ AdminToken: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Ledger entries, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Paginated'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/LedgerEntry' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/store/ledger/reverse:
    post:
      operationId: postAdminStoreLedgerReverse
      tags: [Admin Store]
      summary: Reverse a ledger entry
      description: Posts an opposite `SYSTEM_REVERSAL` entry. Each entry can be reversed once; reversals cannot be reversed. Fails if the customer no longer holds the gold. Reversing a redemption also cancels the request and fails once the gold was collected.
      security: [{ AdminToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ledger_uuid]
              properties:
                ledger_uuid: { type: string, format: uuid }
      responses:
        '200':
          description: Reversed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  message: { type: string }
                  trade: { $ref: '#/components/schemas/LedgerEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Already reversed, not reversible, or redemption no longer pending.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /admin/store/analytics:
    get:
      operationId: getAdminStoreAnalytics
      tags: [Admin Store]
      summary: Tenant trading totals
      security: [{ AdminToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      responses:
        '200':
          description: Totals.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      total_volume_grams: { type: number }
                      total_revenue_inr: { type: number }
                      total_margin_earned: { type: number }
                      total_transactions: { type: integer }
  /admin/store/trade/counter:
    post:
      operationId: postAdminStoreTradeCounter
      tags: [Admin Store]
      summary: Record an in-store (counter) trade
      security: [{ AdminToken: [] }]
      parameters: [{ $ref: '#/components/parameters/TenantId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TradeRequest'
                - type: object
                  required: [user_uuid]
                  properties:
                    user_uuid:
                      {
                        type: string,
                        format: uuid,
                        description: Customer of this tenant.,
                      }
                    payment_mode:
                      {
                        type: string,
                        enum: [COUNTER_CASH, COUNTER_UPI],
                        default: COUNTER_CASH,
                      }
      responses:
        '200':
          description: Trade executed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  trade: { $ref: '#/components/schemas/LedgerEntry' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Slippage exceeded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...